# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NEXUS — API del Sistema Principal
  description: |
    API del módulo `SP`. **Toda operación exige un token de acceso**
    salvo el inicio, la renovación y el cierre de sesión.

    Para probar cualquier endpoint: obtenga el token con
    `POST /api/v1/auth/login`, pulse **Authorize** y pegue el valor de
    `accessToken` — sin el prefijo `Bearer`, que Swagger añade solo.

    El token vive **quince minutos** y no es revocable: cuando expire,
    renuévelo con `POST /api/v1/auth/refresh` y vuelva a autorizar.
  version: v1
servers:
- url: /
  description: Este mismo servidor
security:
- bearerAuth: []
...
paths:
...
  /api/v1/users:
    get:
      tags:
      - Usuarios
      summary: Consultar personas
      description: |
        Listado paginado con filtros, búsqueda y ordenamiento.

        **El orden por defecto es `lastName,asc`** y no el nombre de usuario:
        esta es la lista desde la que se administra el acceso, y quien la mira
        busca a alguien por su apellido.

        Solo se puede ordenar por la lista blanca —`username`, `email`,
        `firstName`, `lastName`, `status`, `createdAt`, `updatedAt`—. **No se
        admite ordenar por ningún campo de la credencial**: ordenar por la marca
        de cambio obligatorio produciría la lista de quien no ha cambiado su
        contraseña inicial.

        **`roleId`, `membershipId` y `countryId` no se validan contra su
        catálogo.** Un filtro por algo inexistente devuelve la colección vacía y
        no es un error.

        **`countryId` NO se acota a países activos**, y va contra la intuición:
        filtrar solo por los activos convertiría desactivar un país en una forma
        de **esconder a su gente**, y este listado es justamente la herramienta
        con la que se busca a quien quedó dentro para moverlo. Quien esté en un
        país retirado aparece con normalidad.

        Cada fila trae su `country` **resuelto y nunca nulo** — es el único
        objeto anidado de la fila del que se puede decir eso: `roles` puede venir
        vacía y `membership` y `deletedAt` pueden venir nulos.

        **El listado NO publica el documento ni los datos de contacto**, y no se
        puede buscar por ellos. Es deliberado: exponer un documento en un listado
        paginado alcanza a mucha más gente que devolverlo en un detalle. Quien
        necesite el documento de una persona usa `GET /api/v1/users/{id}`.

        La búsqueda va sobre nombre de usuario, correo y nombre completo,
        **sin distinguir acentos ni mayúsculas**, y por fragmento.

        `membership` **no es nula cuando está vencida**: vencer no es lo mismo
        que no tener. `current` dice cuál de los dos casos es.

        Un filtro sin coincidencias devuelve `200` con la colección vacía. No
        hay `404` ni `422`.
      operationId: listar
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: sort
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: roleId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: membershipId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: countryId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: includeDeleted
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: Página de personas.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponse"
        "400":
          description: "Paginación fuera de límites (`VAL-001`, `VAL-002`), campo\
            \ de ordenamiento no admitido (`VAL-003`), o estado fuera de su dominio\
            \ (`VAL-004`). Se devuelven **juntos**"
        "401":
          description: Token ausente o inválido (`AUTH-001`)
        "403":
          description: Autenticado sin `users:read` (`AUTH-002`)
        "500":
          description: Fallo no controlado (`ERR-500`)
    post:
      tags:
      - Usuarios
      summary: Registrar una persona
      description: |
        Registra una persona con sus roles iniciales.

        La cuenta nace **activa** y **marcada para cambio obligatorio de
        contraseña**: quien prepara el alta conoce la credencial, y esa ventana
        se cierra en el primer inicio de sesión. Ni el estado ni la marca se
        envían; enviarlos devuelve `400`.
        **`membershipId` es OPCIONAL, y el superior es condicional.** Desde el
        05-09-2026 toda persona nace con nivel: si no se indica ninguno, se le
        concede el de arranque (`BECA`). Ya no hay rechazo por indicarlo sin rol
        de consumidor ni por omitirlo teniéndolo — esa exigencia murió con
        `RN-SP-013`.

        El rol de vendedor y el superior comercial **sí siguen siendo
        condicionales en los dos sentidos**: indicar uno sin el otro devuelve
        `409`, no se ignora.

        El superior debe portar el **rol padre inmediato** del rol vendedor de
        mayor rango de la persona, y estar activo.

        ## Identidad, contacto y país

        **`countryId`, `documentTypeId`, `documentNumber` y `phone` son
        obligatorios**, y a diferencia de la membresía y el superior **no
        dependen de qué roles se concedan**: su ausencia es `400`, nunca un
        `409` condicional.

        **`companyPhone`, `addressLine1`, `addressLine2` y `city` son
        opcionales.** Exigir una dirección postal a un funcionario interno
        bloquearía su alta sin que nadie la necesite. `city` es **texto
        libre**: no hay catálogo de ciudades.

        **El tipo de documento sale de `GET /api/v1/document-types`, y ese
        catálogo contiene SOLO documentos de persona mayor de edad.** Ahí está
        la validación de mayoría de edad del sistema entero: no hay ninguna
        marca que diga cuáles acreditan y cuáles no, porque **los que no
        acreditan no están**. Registrar a un menor no se rechaza — no se puede
        expresar, porque no hay identificador que enviar. Un desplegable
        alimentado por ese catálogo no puede ofrecer una opción que el alta vaya
        a rechazar.

        **El par tipo + número es único entre TODAS las personas, incluidas las
        eliminadas** (`RN-SP-035`), igual que el nombre de usuario y el correo.
        El `409` que produce **no dice de quién es** el documento, ni si esa
        persona sigue vigente.

        El número de documento se guarda **recortado y en mayúsculas**, y
        los **dos teléfonos** —`phone`, el personal, y `companyPhone`, el de la
        empresa— **normalizados a dígitos con un `+` opcional**, fuera espacios,
        guiones y paréntesis. La dirección y la ciudad solo se recortan.

        **`companyPhone` es OPCIONAL y `phone` no** (`RN-SP-037`): exigir un
        teléfono de empresa bloquearía el alta de todo el que no tenga una. Sale
        **presente y en nulo** cuando no se declara, nunca ausente.

        **Ningún teléfono se valida contra el país**: eso exigiría un catálogo de
        prefijos que no existe.

        La respuesta agrupa `country`, `document` y `contact` en tres objetos, y
        devuelve el **tipo de documento resuelto** —abreviación y nombre— para
        que no haga falta una segunda llamada al catálogo, que además exige otro
        permiso.

        La contraseña no se recorta: un espacio al principio o al final es parte
        de ella.
      operationId: registrar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
        required: true
      responses:
        "201":
          description: Persona registrada. `Location` apunta a su detalle.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: "Formato u obligatoriedad incumplidos —incluidos país, tipo\
            \ y número de documento y teléfono, que son obligatorios (`VAL-014` a\
            \ `VAL-017`)—, contraseña que no cumple la política, o campo no admitido\
            \ en el cuerpo"
        "401":
          description: Token ausente o inválido (`AUTH-001`)
        "403":
          description: Autenticado sin el permiso de creación de usuarios (`AUTH-002`)
        "409":
          description: "Identidad ya en uso (`RN-SP-016`), documento ya registrado\
            \ por otra persona —vigente o eliminada— o tipo de documento inactivo\
            \ (`RN-SP-035`), país inactivo (`RN-SP-034`), rol que excede los privilegios\
            \ del actor (`RN-SEG-010`), vendedor sin superior o al revés (`RN-SP-019`),\
            \ o superior que no porta el rol padre (`RN-SP-020`)"
        "422":
          description: "Algún rol no existe o no está activo (`EX-003`), el país no\
            \ existe (`EX-009`) o el tipo de documento no existe (`EX-010`) — referencias\
            \ que no resuelven, frente al `409` de las que resuelven y una regla rechaza"
        "500":
          description: Fallo no controlado (`ERR-500`)
...
  /api/v1/user-commission-rates:
    get:
      tags:
      - Comisiones
      summary: Consultar las tasas personalizadas
      description: |
        Devuelve las tasas **tal como se declararon**, y no resuelve cuál se
        aplica: para eso está `GET /api/v1/commissions/effective`.

        **Incluye el historial**: las vencidas viajan junto a la vigente salvo que
        se filtre por fecha con `onDate`. Y es **el único historial que le queda al
        módulo** — las tasas de rol perdieron la vigencia.

        **Filtrar por persona devuelve las declaradas PARA esa persona**, no la que
        le aplica hoy sobre un producto.

        **Filtrar por producto (`productId`) devuelve las ASOCIADAS a ese producto**,
        de cualquier persona: es la respuesta a «quién tiene excepción aquí». Se
        combina con los demás filtros — persona y producto juntos responden
        «¿tiene esta persona excepción en este producto?», con su historial. Un
        producto donde nadie tiene excepción devuelve la página vacía, y **no
        significa que no comisione**: significa que todos cobran por su rol.

        Cada fila trae **`associatedProducts`**, sobre cuántos productos rige.
        **Cero significa que esa excepción no paga nada** — es el estado en que
        nace, hasta que se asocia. Los productos van contados y no listados; la
        lista está en `GET /api/v1/user-commission-rates/{id}/products`.
      operationId: listar_1
      parameters:
      - name: filtros
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/ListUserCommissionRatesRequest"
      responses:
        "200":
          description: Página de tasas personalizadas
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserCommissionRatePageResponse"
        "400":
          description: Parámetros inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserCommissionRatePageResponse"
        "403":
          description: Sin permiso
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserCommissionRatePageResponse"
    post:
      tags:
      - Comisiones
      summary: Registrar la tasa personalizada de una persona
      description: |
        Declara que **esta persona** gana este porcentaje, y **deja la tasa
        creada sin que rija todavía en ninguna parte**.

        **Para ponerla en vigor hay que asociarla a productos**, con
        `POST /api/v1/user-commission-rates/{id}/products`. Es exactamente el
        mismo mecanismo que las tasas de rol, y desde el 11-09-2026 sin
        excepción: `RN-CM-012` dice que **ninguna tasa rige hasta que se
        asocia**, y esta era la única que se saltaba la regla — pagaba desde el
        primer día sobre **todo el catálogo**.

        **Lo que eso cuesta conviene saberlo**: una tasa creada y no asociada
        **parece configurada y no paga nada**, y eso no falla — se descubre
        liquidando.

        **No lleva rol, y eso tiene una consecuencia que conviene conocer**: la
        tasa sigue rigiendo aunque su titular pase a un rol que no comisiona.
        Hasta el 01-09-2026 el rol era obligatorio precisamente para impedirlo.

        **Aquí NO se comprueba el solapamiento**, y es deliberado: sin producto
        no hay solapamiento posible. Dos tasas de la misma persona con fechas que
        se pisan son legítimas mientras rijan en productos distintos, y
        `RN-CM-006` se verifica **al asociar**.

        **Tampoco se acota el `fixedAmount`**: una tasa sin asociar no conoce el
        precio de nada (`RN-CM-018`). El tope lo pone `RN-CM-019` al asociarla.

        `validFrom` es obligatorio; sin `validTo`, rige indefinidamente.
      operationId: registrar_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserCommissionRateRequest"
        required: true
      responses:
        "201":
          description: Tasa registrada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserCommissionRateResponse"
        "400":
          description: Datos inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserCommissionRateResponse"
        "403":
          description: Sin permiso
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserCommissionRateResponse"
        "422":
          description: La persona no existe
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserCommissionRateResponse"
...
  /api/v1/roles:
    get:
      tags:
      - Roles
      summary: Consultar roles
      description: |
        Listado paginado con filtros, búsqueda y ordenamiento. Es la entrada
        natural a la administración de accesos: de aquí se navega al detalle de
        cada rol.

        **El orden por defecto es `code,asc`**: el código es el identificador
        con el que se habla de un rol en la documentación y en la auditoría.
        Solo se puede ordenar por la lista blanca —`code`, `name`, `roleType`,
        `status`, `createdAt`, `updatedAt`—; cualquier otro campo devuelve
        `400` y **no llega a la base de datos**.

        **No incluye los permisos de cada rol** ni el número de usuarios
        asignados: la primera pregunta la responde el detalle
        (`GET /api/v1/roles/{id}`) y la segunda se hace sobre un rol concreto,
        no sobre la lista.

        **`parentRoleId` no se valida contra el catálogo.** Filtra por el padre
        **directo** —no por el subárbol— y uno inexistente devuelve la
        colección vacía, que no es un error.

        La búsqueda va sobre código y nombre, **sin distinguir acentos ni
        mayúsculas**, y por fragmento: `ministracion` encuentra
        `ADMINISTRACION`. Un término en blanco equivale a no filtrar.

        **Los eliminados quedan fuera salvo que se pidan** con
        `includeDeleted=true`; cuando se piden, `deletedAt` es lo que permite
        distinguirlos.
      operationId: listar_2
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: sort
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: roleType
        in: query
        required: false
        schema:
          type: string
      - name: parentRoleId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: includeDeleted
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: Página de roles.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponse"
        "400":
          description: "Paginación fuera de límites, campo de ordenamiento no admitido\
            \ (`VAL-003`), o `status`/`roleType` fuera de su dominio (`VAL-004`).\
            \ Se evalúan y se devuelven **juntos** en `errors`"
        "401":
          description: Token ausente o inválido (`AUTH-001`)
        "403":
          description: Autenticado sin `roles:read` (`AUTH-002`)
        "500":
          description: Fallo no controlado (`ERR-500`)
    post:
      tags:
      - Roles
      summary: Registrar un rol
      description: |
        Registra un rol bajo un rol padre existente y activo, con los permisos que
        se declaren.

        Los permisos deben estar contenidos en los del rol padre (`RN-SEG-003`) y en
        los permisos efectivos de quien ejecuta el alta (`RN-SEG-010`). Declararlos es
        opcional: sin ellos el rol queda registrado y a la espera de `RF-SP-005`.

        El rol nace siempre **activo** y nunca de sistema. Ni `status` ni `isSystem`
        se admiten en el cuerpo: enviarlos devuelve `400`.
      operationId: registrar_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
        required: true
      responses:
        "201":
          description: Rol registrado. La cabecera `Location` lleva su dirección.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RoleResponse"
        "400":
          description: "Formato u obligatoriedad incumplidos (`VAL-001` a `VAL-004`,\
            \ `VAL-007`, `VAL-008`), o el cuerpo trae un campo que este endpoint no\
            \ admite"
        "401":
          description: Token ausente o inválido (`AUTH-001`)
        "403":
          description: Autenticado sin `roles:create` (`AUTH-002`)
        "409":
          description: "Código o nombre ya en uso (`RN-SEG-001`), permiso fuera del\
            \ rol padre (`RN-SEG-003`) o fuera del alcance del actor (`RN-SEG-010`)"
        "422":
          description: "El rol padre no existe o no está activo (`EX-002`), o algú\
            n permiso no está en el catálogo (`EX-005`)"
        "500":
          description: Fallo no controlado (`ERR-500`)
...
  /api/v1/movements:
    post:
      tags:
      - Movimientos
      summary: Registrar una venta a nombre de un cliente
      description: |
        Deja constancia de **qué le vendió la empresa a un cliente**, como un hecho que
        **todavía no está pagado**.

        **La venta nace `PENDIENTE`, y eso significa que no concede nada.** No sube de
        nivel a nadie, no habilita ninguna cuenta y no comisiona: registrar una venta
        **no cambia absolutamente nada fuera de este módulo**. Confirmarla es otra
        operación.

        **El precio no se envía: se toma del catálogo.** Se indica qué productos y
        cuántos, nunca cuánto cuestan — un precio que llegara en la petición sería un
        descuento sin autorización y sin rastro. Tampoco se envían la moneda, la
        vigencia ni el vendedor: la moneda y la vigencia salen del producto, y **el
        vendedor sale del cliente** y se congela en la venta.

        **Lo copiado queda congelado.** Corregir mañana el precio de un producto, o
        reasignar el cliente a otro agente, no cambia lo que se vendió hoy.

        Reglas de composición: **como mucho un upgrade** por venta y con cantidad uno,
        sin productos repetidos y todas las líneas en la misma moneda.
      operationId: registrar_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterSaleRequest"
        required: true
      responses:
        "201":
          description: "Venta registrada, pendiente de pago."
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SaleResponse"
        "400":
          description: "Lo que se ve mirando la petición: falta el cliente o el mé\
            todo de pago, no hay líneas, la cantidad no es positiva, un producto se\
            \ repite, o la fecha del hecho está en el futuro."
        "403":
          description: Sin el permiso `movements:create`.
        "409":
          description: "Lo que solo se sabe después de resolver: la cuenta no puede\
            \ operar todavía, el cliente no cuelga de ningún vendedor, un producto\
            \ no está en su oferta, el upgrade BAJA de nivel —renovar el mismo sí\
            \ se admite—, hay dos upgrades, las monedas difieren, o el método de pago\
            \ está desactivado."
        "422":
          description: "Un dato bien formado que no resuelve: el cliente, un producto\
            \ o el método de pago no existen."
...
  /api/v1/movements/mine:
    get:
      tags:
      - Movimientos
      summary: Consultar los movimientos propios
      description: |
        Devuelve **los movimientos en los que usted participó**, paginados y del más
        reciente al más antiguo.

        **«Propio» son DOS papeles.** Un movimiento lleva a quien recibe lo comprado y a
        quien lo vendió, y usted puede ser cualquiera de los dos — o **los dos a la vez**,
        si compró para sí mismo algo que se le atribuye. Cada movimiento dice en qué papel
        aparece usted con `role`: `BUYER`, `SELLER` o `BOTH`. El que es las dos cosas
        **aparece una sola vez**.

        **No hay forma de preguntar por otra persona**, ni indicándola ni teniendo
        permisos: quien pregunta sale de la credencial. Consultar las ventas de terceros es
        otra operación, con su permiso.

        **Las líneas no viajan aquí.** Una venta puede llevar varias, y meterlas
        multiplicaría la respuesta por un dato que solo se mira al abrir uno: están en el
        detalle.

        **El vendedor puede ser nulo**, y viaja igual: es el caso normal de quien no cuelga
        de ningún vendedor.

        El orden es fijo y no se puede cambiar. `status` filtra por estado.
      operationId: mios
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: La página de movimientos propios.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseMyMovement"
        "400":
          description: Paginación inválida (`VAL-002`) o estado no admitido (`VAL-003`)
        "401":
          description: Token ausente o inválido (`AUTH-001`)
        "500":
          description: Fallo no controlado (`ERR-500`)
...
components:
  schemas:
...
    RoleRef:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        status:
          type: string
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
        mustChangePassword:
          type: boolean
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleRef"
        country:
          $ref: "#/components/schemas/CountryRef"
        document:
          $ref: "#/components/schemas/DocumentRef"
        contact:
          $ref: "#/components/schemas/ContactRef"
        membership:
          $ref: "#/components/schemas/MembershipRef"
        supervisor:
          $ref: "#/components/schemas/SupervisorRef"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PageResponse:
      type: object
      properties:
        content:
          type: array
          items: {}
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalIsExact:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      description: "Token de acceso emitido por `POST /api/v1/auth/login`. Pegue solo\
        \ el valor de `accessToken`: el prefijo `Bearer` lo añade la interfaz."
      in: header
      scheme: bearer
      bearerFormat: JWT